pyproject.toml
[tool.poetry]
package-mode = true
name = "erkg-tutorials-graphgeeks"
version = "0.1.0"
description = "Tutorials for Entity Resolved Knowledge Graphs, doing Entity Linking on ICIJ articles."
authors = ["Louis Guitton <louisguitton93@gmail.com>", "Paco Nathan <paco@derwen.ai>"]
readme = "README.md"
homepage = "https://guitton.co/"
repository = "https://github.com/louisguitton/erkg-tutorials"
documentation = "https://guitton.co/"
keywords = ["spacy", "senzing", "entity-linking", "icij"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
spacy-lancedb-linker = "^0.1.2"
requests = "^2.32.3"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pandas = "^2.2.3"

[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"
# https://python-poetry.org/docs/dependency-specification/#url-dependencies
en_core_web_md = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz" }
# notebook environment
icecream = "^2.1"
ipywidgets = "^8.1"
jupyterlab = "^4.2"
jupyterlab_execute_time = "^3.1"
watermark = "^2.4"
pandas-stubs = "^2.2.2.240909"
types-tqdm = "^4.66.0.20240417"
types-requests = "^2.32.0.20240914"
types-beautifulsoup4 = "^4.12.0.20240907"
dagster = "^1.8.11"
dagster-webserver = "^1.8.11"

[tool.poetry.scripts]
senzing-pipeline = "src.senzing_pipeline:main"
scraper = "src.scraper:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black#pyprojecttoml
line-length = 100


[tool.isort]
# https://github.com/timothycrosley/isort#configuring-isort
profile = "black"

[tool.dagster]
module_name = "dagster_icij.definitions"
code_location_name = "dagster_icij"
